requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.4
fastapi-cli==0.0.5
frontend==0.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.1
Pillow==9.0.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.13
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
pyzbar==0.1.9
requests==2.32.3
rich==13.9.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
